+++ phone-test/src/pages/CallsList.tsx
@@ -2,17 +2,13 @@ import { useQuery } from '@apollo/client';
 import styled from 'styled-components';
 import { PAGINATED_CALLS } from '../gql/queries';
 import {
-  Grid,
-  Icon,
   Typography,
   Spacer,
   Box,
-  DiagonalDownOutlined,
-  DiagonalUpOutlined,
   Pagination
 } from '@aircall/tractor';
-import { formatDate, formatDuration } from '../helpers/dates';
 import { useNavigate, useSearchParams } from 'react-router-dom';
+import CallListItem from '../components/calls/CallListItem/CallListItem';
 
 export const PaginationWrapper = styled.div`
   > div {
@@ -46,13 +42,11 @@ export const CallsListPage = () => {
 
   const { totalCount, nodes: calls } = data?.paginatedCalls;
 
-  const handleCallOnClick = (callId: string) => {
-    navigate(`/calls/${callId}`);
-  };
-
   const handlePageChange = (page: number) => {
-    console.log(page, Math.ceil(totalCount / callsPerPage), totalCount, callsPerPage)
-    if (page === activePage || page < 1 || page > Math.ceil(totalCount / callsPerPage)) return;
+    console.log(page, Math.ceil(totalCount / callsPerPage), totalCount, callsPerPage);
+    if (page === activePage) return;
+    if (page < 1) page= 1;
+    if (page > Math.ceil(totalCount / callsPerPage)) page= Math.ceil(totalCount / callsPerPage);
     setSearchParams(params => {
       params.set('page', page.toString());
       return params;
@@ -75,59 +69,13 @@ export const CallsListPage = () => {
       <Typography variant="displayM" textAlign="center" py={3}>
         Calls History
       </Typography>
-      <Spacer space={3} direction="vertical">
-        {calls.map((call: Call) => {
-          const icon = call.direction === 'inbound' ? DiagonalDownOutlined : DiagonalUpOutlined;
-          const title =
-            call.call_type === 'missed'
-              ? 'Missed call'
-              : call.call_type === 'answered'
-              ? 'Call answered'
-              : 'Voicemail';
-          const subtitle = call.direction === 'inbound' ? `from ${call.from}` : `to ${call.to}`;
-          const duration = formatDuration(call.duration / 1000);
-          const date = formatDate(call.created_at);
-          const notes = call.notes ? `Call has ${call.notes.length} notes` : <></>;
-
-          return (
-            <Box
-              key={call.id}
-              bg="black-a30"
-              borderRadius={16}
-              cursor="pointer"
-              onClick={() => handleCallOnClick(call.id)}
-            >
-              <Grid
-                gridTemplateColumns="32px 1fr max-content"
-                columnGap={2}
-                borderBottom="1px solid"
-                borderBottomColor="neutral-700"
-                alignItems="center"
-                px={4}
-                py={2}
-              >
-                <Box>
-                  <Icon component={icon} size={32} />
-                </Box>
-                <Box>
-                  <Typography variant="body">{title}</Typography>
-                  <Typography variant="body2">{subtitle}</Typography>
-                </Box>
-                <Box>
-                  <Typography variant="caption" textAlign="right">
-                    {duration}
-                  </Typography>
-                  <Typography variant="caption">{date}</Typography>
-                </Box>
-              </Grid>
-              <Box px={4} py={2}>
-                <Typography variant="caption">{notes}</Typography>
-              </Box>
-            </Box>
-          );
-        })}
-      </Spacer>
-
+      <Box overflow="auto" minWidth={500} paddingTop={5} maxHeight="80vh">
+        <Spacer space={3} direction="vertical">
+          {calls.map((call: Call) => (
+            <CallListItem call={call} key={call.id} />
+          ))}
+        </Spacer>
+      </Box>
       {totalCount && (
         <PaginationWrapper>
           <Pagination
